Rename archive files for clarity
